refactor(frame): 10프레임 구현을 위해서 Frame을 인터페이스로 분리
